+++ Packages/ai.fxn.fxn3d/Runtime/Function.cs
@@ -96,7 +96,7 @@ namespace Function {
 
         #region --Operations--
         public readonly FunctionClient client;
-        public const string Version = @"0.0.34";
+        public const string Version = @"0.0.35";
         internal const string URL = @"https://api.fxn.ai/v1";
         #endregion
     }